refactor: Move data rage button group to parent
Move button group from VaultChart to VaultDetailsTabWrapper for easier separation of Overview and VaultChart
